Take projectile per shot into account when comparing TTK and BTK
First Pass on variants with attachments. Added the variant count of a specific weapon as the prefix of the weapons name and attachments. variant count/id is used to filter correct row when updating weapons.

Removed js script from recoil comparision HTML. The figure object is now called on page initialization.